object mapping working, but not good

+++ prob_map_view/include/prob_map_view/ProbViewer.h
@@ -12,9 +12,9 @@
 
 class ProbViewer : public ProbViewer_B{
   protected:
-    std::vector<cv::Mat_<float>> prob_images_;
+    std::vector<cv::Mat_<double>> prob_images_;
     std::vector<std::string> prob_names_;
-    cv::Mat_<float> curr_img_;
+    cv::Mat_<double> curr_img_;
     bool img_are_log_;
     wxChoice* select_choice_;
 
@@ -25,7 +25,7 @@ class ProbViewer : public ProbViewer_B{
     virtual void onCheck( wxCommandEvent& event );
     virtual void onMouseMove( wxMouseEvent& event );
 
-    void updateImages(const std::vector<cv::Mat_<float>> prob_images);
+    void updateImages(const std::vector<cv::Mat_<double>> prob_images);
     void setCurrent();
 };
